Removed unneccessary project directory, commented most of project
