--- .idea/compiler.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="CompilerConfiguration">
-    <wildcardResourcePatterns>
-      <entry name="!?*.java" />
-      <entry name="!?*.form" />
-      <entry name="!?*.class" />
-      <entry name="!?*.groovy" />
-      <entry name="!?*.scala" />
-      <entry name="!?*.flex" />
-      <entry name="!?*.kt" />
-      <entry name="!?*.clj" />
-    </wildcardResourcePatterns>
-    <bytecodeTargetLevel>
-      <module name="CrabSim.main" target="1.8" />
-      <module name="CrabSim.test" target="1.8" />
-    </bytecodeTargetLevel>
-  </component>
-</project>
