--- .idea/libraries/Gradle__org_jetbrains_annotations_16_0_2.xml
@@ -1,11 +1,0 @@
-<component name="libraryTable">
-  <library name="Gradle: org.jetbrains:annotations:16.0.2">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.jetbrains/annotations/16.0.2/dab5655eb9abe23f2245c34e66c9813c7daf4563/annotations-16.0.2.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.jetbrains/annotations/16.0.2/11ef7b63fea95edbd3ea28f564d8bec5d86b5d0e/annotations-16.0.2-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
