--- .idea/modules/CrabSim.iml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module external.linked.project.id="CrabSim" external.linked.project.path="$MODULE_DIR$/../../CrabSim" external.root.project.path="$MODULE_DIR$/../../CrabSim" external.system.id="GRADLE" external.system.module.group="" external.system.module.version="1.0-SNAPSHOT" type="JAVA_MODULE" version="4">
-  <component name="ExternalSystem" linkedProjectPath="$MODULE_DIR$/../../CrabSim" rootProjectPath="$MODULE_DIR$/../../CrabSim" />
-  <component name="NewModuleRootManager" inherit-compiler-output="true">
-    <exclude-output />
-    <content url="file://$MODULE_DIR$/../../CrabSim">
-      <excludeFolder url="file://$MODULE_DIR$/../../CrabSim/.gradle" />
-      <excludeFolder url="file://$MODULE_DIR$/../../CrabSim/build" />
-    </content>
-    <orderEntry type="inheritedJdk" />
-    <orderEntry type="sourceFolder" forTests="false" />
-  </component>
-</module>
