--- .idea/modules/CrabSim.test.iml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module external.linked.project.id="CrabSim:test" external.linked.project.path="$MODULE_DIR$/../../CrabSim" external.root.project.path="$MODULE_DIR$/../../CrabSim" external.system.id="GRADLE" external.system.module.group="" external.system.module.type="sourceSet" external.system.module.version="1.0-SNAPSHOT" type="JAVA_MODULE" version="4">
-  <component name="ExternalSystem" linkedProjectPath="$MODULE_DIR$/../../CrabSim" rootProjectPath="$MODULE_DIR$/../../CrabSim" />
-  <component name="NewModuleRootManager" LANGUAGE_LEVEL="JDK_1_8">
-    <output-test url="file://$MODULE_DIR$/../../CrabSim/build/classes/java/test" />
-    <exclude-output />
-    <content url="file://$MODULE_DIR$/../../CrabSim/src/test">
-      <sourceFolder url="file://$MODULE_DIR$/../../CrabSim/src/test/java" isTestSource="true" />
-      <sourceFolder url="file://$MODULE_DIR$/../../CrabSim/src/test/resources" type="java-test-resource" />
-    </content>
-    <orderEntry type="inheritedJdk" />
-    <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="module" module-name="CrabSim.main" />
-    <orderEntry type="library" name="Gradle: org.ini4j:ini4j:0.5.1" level="project" />
-    <orderEntry type="library" name="Gradle: org.jetbrains:annotations:16.0.2" level="project" />
-    <orderEntry type="library" name="Gradle: junit:junit:4.12" level="project" />
-    <orderEntry type="library" name="Gradle: org.hamcrest:hamcrest-core:1.3" level="project" />
-  </component>
-  <component name="TestModuleProperties" production-module="CrabSim.main" />
-</module>
