--- CrabSim/src/main/configuration.ini
@@ -1,12 +1,0 @@
-[Meta]
-fullscreen=true
-binRegionSize=10
-
-[Overworld]
-squareRegionSet=true
-horizontalRegions=2
-verticalRegions=2
-basictile=FlatTerrain
-
-[Enemies]
-numOfLobsters=100
