--- CrabSim/src/main/java/entities/enemies/Pelican.java
@@ -1,4 +1,0 @@
-package entities.enemies;
-
-public class Pelican {
-}
