--- CrabSim/src/main/java/metaControl/KeyAction.java
@@ -1,5 +1,0 @@
-package metaControl;
-
-public interface KeyAction {
-    void action();
-}
