+++ src/main/java/entities/enemies/Enemy.java
@@ -2,5 +2,8 @@ package entities.enemies;
 
 import entities.Being;
 
+/**
+ * Defines and distinguishes an enemy from a plain being.
+ */
 public abstract class Enemy extends Being {
 }
