+++ src/main/java/entities/enemies/Pelican.java
@@ -1,0 +1,7 @@
+package entities.enemies;
+
+/**
+ * The Pelican moves at a fast rate of 8 but does not shoot.
+ */
+public class Pelican {
+}
