+++ src/main/java/entities/players/Crab.java
@@ -10,6 +10,9 @@ import javafx.util.Duration;
 import main.AssetLoading;
 import main.Main;
 
+/**
+ * The Crab player moves at an average speed of 10 and shoots at an average rate of UNDEFINED.
+ */
 public class Crab extends Player {
 
     private final static int CRAB_SPEED = 10;
@@ -27,6 +30,10 @@ public class Crab extends Player {
         });
     }
 
+    /**
+     * @inheritDoc
+     * @param damageAmount The damage amount to be dealt
+     */
     @Override
     public void takeDamage(int damageAmount) {
         health -= damageAmount;
@@ -57,69 +64,108 @@ public class Crab extends Player {
         }
     }
 
+    /**
+     * @inheritDoc
+     */
     @Override
     void toggleVulnerability() {
         vulnerable = !vulnerable;
     }
 
+    /**
+     * @inheritDoc
+     */
     @Override
     public void setRotateLeft() {
         rotatingLeft = true;
         rotateLeft();
     }
 
+    /**
+     * @inheritDoc
+     */
     @Override
     public void setRotateRight() {
         rotatingRight = true;
         rotateRight();
     }
 
+    /**
+     * @inheritDoc
+     */
     @Override
     public void setShoot() {
         shooting = true;
     }
 
+    /**
+     * @inheritDoc
+     */
     @Override
     public void rotateLeft() {
         this.setRotate((Math.abs((int) this.getRotate()) % 360 <= 1) ? -(Main.deltaTime * CRAB_SPEED) : this.getRotate() - (Main.deltaTime * CRAB_SPEED));
     }
 
+    /**
+     * @inheritDoc
+     */
     @Override
     public void rotateRight() {
         this.setRotate((Math.abs((int) this.getRotate()) % 360 <= 1) ? (Main.deltaTime * CRAB_SPEED) : this.getRotate() + (Main.deltaTime * CRAB_SPEED));
     }
 
+    /**
+     * @inheritDoc
+     */
     @Override
     public void stopRightRotate() {
         rotatingRight = false;
     }
 
+    /**
+     * @inheritDoc
+     */
     @Override
     public void stopLeftRotate() {
         rotatingLeft = false;
     }
 
+    /**
+     * @inheritDoc
+     */
     @Override
     void moveForward() {
         this.setTranslateX(this.getTranslateX() + ( ( ((CRAB_SPEED * 5) * Main.deltaTime) * Math.sin(Math.toRadians(-(this.getRotate() - 90))) ) ) );
         this.setTranslateY(this.getTranslateY() + ( ( ((CRAB_SPEED * 5) * Main.deltaTime) * Math.cos(Math.toRadians(-(this.getRotate() - 90))) ) ) );
     }
 
+    /**
+     * @inheritDoc
+     */
     @Override
     public void shoot() {
 
     }
 
+    /**
+     * @inheritDoc
+     */
     @Override
     public void stopShoot() {
 
     }
 
+    /**
+     * @inheritDoc
+     */
     @Override
     protected Rectangle getBoundingBox() {
         return new Rectangle(this.getLayoutX() + this.getTranslateX(), this.getLayoutY() + this.getTranslateY(),100, 50);
     }
 
+    /**
+     * @inheritDoc
+     */
     @Override
     public void update() {
         if (rotatingLeft)
@@ -132,16 +178,25 @@ public class Crab extends Player {
             moveForward();
     }
 
+    /**
+     * @inheritDoc
+     */
     @Override
     protected boolean isVulnerable() {
         return vulnerable;
     }
 
+    /**
+     * @inheritDoc
+     */
     @Override
-    public int getDamageAmount() {
+    public int getCollisionDamageAmount() {
         return 0;
     }
 
+    /**
+     * @inheritDoc
+     */
     @Override
     public void kill() {
         EntityHandler.getEntities().remove(this);
@@ -151,6 +206,9 @@ public class Crab extends Player {
         stopMovement();
     }
 
+    /**
+     * @inheritDoc
+     */
     @Override
     public void stopMovement() {
         moving = false;
