+++ src/main/java/main/AssetLoading.java
@@ -5,6 +5,9 @@ import javafx.scene.image.Image;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 
+/**
+ * Loads in image assets from the local file system.
+ */
 public class AssetLoading {
 
     public static Image crabSprite;
@@ -22,8 +25,8 @@ public class AssetLoading {
 
     private static void loadPlayers() {
         try {
-            crabSprite = new Image(new FileInputStream("CrabSim/src/main/resources/Players/Crab.png"),0, 0, true, false);
-            shrimpSprite = new Image(new FileInputStream("CrabSim/src/main/resources/Players/Shrimp.png"),0, 0, true, false);
+            crabSprite = new Image(new FileInputStream("src/main/resources/Players/Crab.png"),0, 0, true, false);
+            shrimpSprite = new Image(new FileInputStream("src/main/resources/Players/Shrimp.png"),0, 0, true, false);
         }
         catch (FileNotFoundException error) {
             error.printStackTrace();
@@ -32,7 +35,7 @@ public class AssetLoading {
 
     private static void loadEnemies() {
         try {
-            lobsterSprite = new Image(new FileInputStream("CrabSim/src/main/resources/Enemies/Lobster.png"),0, 0, true, false);
+            lobsterSprite = new Image(new FileInputStream("src/main/resources/Enemies/Lobster.png"),0, 0, true, false);
         }
         catch (FileNotFoundException error) {
             error.printStackTrace();
@@ -41,7 +44,7 @@ public class AssetLoading {
 
     private static void loadTiles() {
         try {
-            defaultWorldTile = new Image(new FileInputStream("CrabSim/src/main/resources/Tiles/"
+            defaultWorldTile = new Image(new FileInputStream("src/main/resources/Tiles/"
                     + LoadConfiguration.getBasicWorldTile() + ".png"),0, 0, true, false);
         }
         catch (FileNotFoundException error) {
