+++ src/main/java/metaControl/CameraControl.java
