+++ src/main/java/splitScreen/SplitScreen.java
