+++ src/main/java/splitScreen/SplitScreenHandler.java
